Changes zoom/lat/lng to be mutable properties
Fixes #284. Using this.el.setAttribute was good for reflecting the
attributes back to the DOM but in polyfilled browsers caused an infinite
feedback loop where the attribute change happened outside of the
"mapMoveInProgress" guard.

Also adds test coverage for map zooming in the UI being reflected in the
element's properties.