Changes the Stencil namespace to "FleetComponents"
The old name of "all" was too likely to be a naming conflict with other
code. For example, the Stencil loader defines a global object with the
namespace name. "FleetComponents" is a better choice for that than
"all".

Because the sync to S3 is additive, we can deploy this and then switch
the sites that link to the web components over later.