Dependency updates
  - Gets us off our forks of Esri libraries and Leaflet
  - Updates Stencil

Note: We pin to Leaflet 1.3.2 exactly because that's the only version
that points to its ESM src explicitly with the "module" property in
package.json.

I was unable to use Rollup / TypeScript module aliases to point
"leaflet" at the leaflet-src.esm.js file directly. It is likely
something with the Stencil config was interfering.

Fixes #253