Adds a mn-fixed-child class for nav menu
Necessary to set the left position of fixed children that should move
out of the way of the side hamburger menu. While not setting a left
position for a fixed element should make it start at the mn element’s
left, Safari does not correctly re-position the element when the value
of left is animated.

See CityOfBoston/311#928